Refactor domain package name
