+++ domain/src/main/java/com/snekbyte/starter/domain/interactors/types/ObservableInteractorWithParameter.java
@@ -14,10 +14,14 @@
  * limitations under the License.
  */
 
-package com.snekbyte.starter.interactors.types;
+package com.snekbyte.starter.domain.interactors.types;
 
-import io.reactivex.Single;
+import io.reactivex.Observable;
 
-public interface SingleInteractor<T> {
-  Single<T> execute();
+/**
+ * @param <P> Parameter object
+ * @param <R> Response data type for the Observable
+ */
+public interface ObservableInteractorWithParameter<P, R> {
+  Observable<R> execute(P param);
 }
